+++ src/core/requests/authentication.py
@@ -160,6 +160,9 @@ def http_auth_cracker(url, realm):
     for username in usernames:
       for password in passwords:
         float_percent = "{0:.1f}%".format(round(((i*100)/(total*1.0)),2))
+        # Check if verbose mode on
+        if menu.options.verbose:
+            sys.stdout.write(Fore.GREY + "(~) Checking: " + username + ":" + password + Style.RESET_ALL + "\n")
         try:
           # Basic authentication 
           if authentication_type.lower() == "basic":
@@ -182,19 +185,23 @@ def http_auth_cracker(url, realm):
           raise 
         except:
           pass  
-
         if found:
-          float_percent = Fore.GREEN + "SUCCEED" + Style.RESET_ALL
+          if not menu.options.verbose:
+            float_percent = Fore.GREEN + "SUCCEED" + Style.RESET_ALL
         else:
           if str(float_percent) == "100.0%":
-            float_percent = Fore.RED + "FAILED" + Style.RESET_ALL
+            if not menu.options.verbose:
+              float_percent = Fore.RED + "FAILED" + Style.RESET_ALL
           else:  
             i = i + 1
-        sys.stdout.write("\r\r" + settings.INFO_SIGN + "Checking for a valid pair of credentials... [ " +  float_percent + " ]")
-        sys.stdout.flush()
+        if not menu.options.verbose:
+          sys.stdout.write("\r\r" + settings.INFO_SIGN + "Checking for a valid pair of credentials... [ " +  float_percent + " ]")
+          sys.stdout.flush()
         if found:
           valid_pair =  "" + username + ":" + password + ""
-          print Style.BRIGHT + "\n(!) Identified a valid pair of credentials '" + Style.UNDERLINE  + valid_pair + Style.RESET_ALL + Style.BRIGHT  + "'." + Style.RESET_ALL
+          if not menu.options.verbose:
+            print ""
+          print Style.BRIGHT + "(!) Identified a valid pair of credentials '" + Style.UNDERLINE  + valid_pair + Style.RESET_ALL + Style.BRIGHT  + "'." + Style.RESET_ALL
           return valid_pair
 
     error_msg = "Use the '--auth-cred' option to provide a valid pair of "