+++ readme/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Version 0.7b
+* Added: Support for verbose mode in HTTP authentication (Basic / Digest) dictionary-based cracker.
 * Added: The ability to store valid (Digest) credentials into session files for current target.
 * Added: Dictionary-based cracker for "Digest" HTTP authentication credentials.
 * Added: Support for "Digest" HTTP authentication type.